Use --single-transaction for mysqldump
InnoDB is an MVCC storage engine. This means it does not need to lock
tables in order to do a consistent read of data. Add
--single-transaction to mysqldump invocations to enable this behavior
and stop locking the tables.

Downsides to this - changes made to MyISAM tables may not be consistent
(but for the love of all that is holy we should be storing nothing in
MyISAM and neither should anyone else) and DDL commands will not be
allowed during the backup.

Change-Id: If2ea9e5a25a3cfe30c16cc646463afbd3b2a5ab5